requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.30.0
python-dotenv>=1.0.0
langgraph>=0.2.0
tavily-python>=0.3.0
pyyaml>=6.0.0
